chore(Domain, Application): Section관리 영역 변경
- Section 단독 -> Line 하위로 관리영역 변경
