+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -2,6 +2,8 @@ package nextstep.subway.line.application;
 
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
+import nextstep.subway.line.domain.Section;
+import nextstep.subway.line.domain.SectionRepository;
 import nextstep.subway.line.dto.LineInfoResponse;
 
 import java.util.List;
@@ -14,9 +16,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class LineService {
     private LineRepository lineRepository;
+    private SectionRepository sectionRepository;
 
-    public LineService(LineRepository lineRepository) {
+    public LineService(LineRepository lineRepository, SectionRepository sectionRepository) {
         this.lineRepository = lineRepository;
+        this.sectionRepository = sectionRepository;
     }
 
     @Transactional
@@ -48,5 +52,10 @@ public class LineService {
     @Transactional
     public void deleteLineInfo(Long lineId) {
         lineRepository.deleteById(lineId);
+    }
+
+    @Transactional
+    public Section saveSection(Section section) {
+        return sectionRepository.save(section);
     }
 }
