+++ src/test/java/nextstep/subway/line/domain/SectionTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.section.domain;
+package nextstep.subway.line.domain;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
